# examples/petstore3-server-go-echo/server/openapi_generated.yml
openapi: 3.0.2
info:
    contact:
        email: apiteam@swagger.io
    description: |-
        This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
        Swagger at [http://swagger.io](http://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
        You can now help us improve the API whether it's by making changes to the definition itself or to the code.
        That way, with time, we can improve the API in general, and expose some of the new features in OAS3.
    
        Some useful links:
        - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
        - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.html
    termsOfService: http://swagger.io/terms/
    title: Swagger Petstore - OpenAPI 3.0
    version: 1.0.19
...
servers:
    - url: /api/v3
...
paths:
    /pet:
        post:
            description: Add a new pet to the store
            operationId: addPet
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreatePet'
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/CreatePet'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/CreatePet'
                description: Create a new pet in the store
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Pet'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Pet'
                    description: Successful operation
                "405":
                    description: Invalid input
            security:
                - petstore_auth:
                    - write:pets
                    - read:pets
            summary: Add a new pet to the store
            tags:
                - pet
        put:
            description: Update an existing pet by Id
            operationId: updatePet
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Pet'
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/Pet'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/Pet'
                description: Update an existent pet in the store
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Pet'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Pet'
                    description: Successful operation
                "400":
                    description: Invalid ID supplied
                "404":
                    description: Pet not found
                "405":
                    description: Validation exception
            security:
                - petstore_auth:
                    - write:pets
                    - read:pets
            summary: Update an existing pet
            tags:
                - pet
    /pet/findByStatus:
        get:
            description: Multiple status values can be provided with comma separated strings
            operationId: findPetsByStatus
            parameters:
                - description: Status values that need to be considered for filter
                  explode: true
                  in: query
                  name: status
                  required: false
                  schema:
                    default: available
                    enum:
                        - available
                        - pending
                        - sold
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Pet'
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Pet'
                                type: array
                    description: successful operation
                "400":
                    description: Invalid status value
            security:
                - {}
                - petstore_auth:
                    - write:pets
                    - read:pets
            summary: Finds Pets by status
            tags:
                - pet
    /pet/findByTags:
        get:
            description: Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.
            operationId: findPetsByTags
            parameters:
                - description: Tags to filter by
                  explode: true
                  in: query
                  name: tags
                  required: false
                  schema:
                    items:
                        type: string
                    type: array
                - $ref: '#/components/parameters/LimitQuery'
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Pet'
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Pet'
                                type: array
                    description: successful operation
                "400":
                    description: Invalid tag value
            security:
                - petstore_auth:
                    - write:pets
                    - read:pets
            summary: Finds Pets by tags
            tags:
                - pet
    /pet/{petId}:
        get:
            description: Returns a single pet
            operationId: getPetById
            parameters:
                - description: ID of pet to return
                  in: path
                  name: petId
                  required: true
                  schema:
                    format: int64
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Pet'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Pet'
                    description: successful operation
                "400":
                    description: Invalid ID supplied
                "404":
                    description: Pet not found
            security:
                - api_key: []
                - petstore_auth:
                    - write:pets
                    - read:pets
            summary: Find pet by ID
            tags:
                - pet
        post:
            description: ""
            operationId: updatePetWithForm
            parameters:
                - description: ID of pet that needs to be updated
                  in: path
                  name: petId
                  required: true
                  schema:
                    format: int64
                    type: integer
                - description: Name of pet that needs to be updated
                  in: query
                  name: name
                  schema:
                    type: string
                - description: Status of pet that needs to be updated
                  in: query
                  name: status
                  schema:
                    type: string
            responses:
                "405":
                    description: Invalid input
            security:
                - petstore_auth:
                    - write:pets
                    - read:pets
            summary: Updates a pet in the store with form data
            tags:
                - pet
        delete:
            description: ""
            operationId: deletePet
            parameters:
                - description: ""
                  in: header
                  name: X-Api-Key
                  required: true
                  schema:
                    type: string
                - description: Pet id to delete
                  in: path
                  name: petId
                  required: true
                  schema:
                    format: int64
                    type: integer
            responses:
                "400":
                    description: Invalid pet value
            security:
                - petstore_auth:
                    - write:pets
                    - read:pets
            summary: Deletes a pet
            tags:
                - pet
    /pet/{petId}/uploadImage:
        post:
            description: ""
            operationId: uploadFile
            parameters:
                - description: ID of pet to update
                  in: path
                  name: petId
                  required: true
                  schema:
                    format: int64
                    type: integer
                - description: Additional Metadata
                  in: query
                  name: additionalMetadata
                  required: false
                  schema:
                    type: string
            requestBody:
                content:
                    application/octet-stream:
                        schema:
                            format: binary
                            type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiResponse'
                    description: successful operation
            security:
                - petstore_auth:
                    - write:pets
                    - read:pets
            summary: uploads an image
            tags:
                - pet
...
components:
    parameters:
        LimitQuery:
            in: query
            name: limit
            schema:
                maximum: 1000
                type: integer
...
    schemas:
...
        ApiResponse:
            properties:
                code:
                    format: int32
                    type: integer
                message:
                    type: string
                type:
                    type: string
            type: object
            xml:
                name: '##default'
...
        Pet:
            allOf:
                - properties:
                    id:
                        example: 10
                        exclusiveMinimum: 0
                        format: int64
                        type: integer
                  required:
                    - id
                  type: object
                - $ref: '#/components/schemas/CreatePet'
            xml:
                name: pet
...
    securitySchemes:
        api_key:
            in: header
            name: api_key
            type: apiKey
        petstore_auth:
            flows:
                implicit:
                    authorizationUrl: https://petstore3.swagger.io/oauth/authorize
                    scopes:
                        read:pets: read your pets
                        write:pets: modify pets in your account
            type: oauth2